# Improved performance with less memory access