--- a/prepare_abeobk.sh
+++ b/prepare_abeobk.sh
@@ -16,5 +16,5 @@
 #
 
 # Uncomment below to use sdk
-# source "$HOME/.sdkman/bin/sdkman-init.sh"
-# sdk use java 21.0.1-graal 1>&2
+source "$HOME/.sdkman/bin/sdkman-init.sh"
+sdk use java 21.0.1-graal 1>&2