--- a/src/main/java/dev/morling/onebrc/CalculateAverage_abeobk.java
+++ b/src/main/java/dev/morling/onebrc/CalculateAverage_abeobk.java
@@ -192,6 +192,7 @@ public class CalculateAverage_abeobk {
                         addr += semi_pos;
 
                         int hash32 = xxh32(hash);
+
                         long keylen = (addr - row_addr);
                         tail = tail | (keylen << 56);